Refactored account creation code into blueprint
- Everything is handled under blueprint registered for /account.
- Blueprint should be used for everything dealing with accounts.
