+++ RuddockWebsite/auth_utils.py
@@ -347,46 +347,3 @@ def check_create_account_key(key):
     return result['user_id']
   else:
     return None
-
-def handle_create_account(user_id, username, password, password2):
-  '''
-  Creates a new account. Flashes a message and returns False if an error
-  occurs, otherwise returns True.
-  '''
-
-  # Validate username and password. The validate_* functions will flash errors.
-  if validate_username(username) and validate_password(password, password2):
-    # Insert new values into the database. Because the password is updated in a
-    # separate step, we must use a transaction to execute this query.
-    trans = g.db.begin()
-    try:
-      # Insert the new row into users.
-      query = text("""
-        INSERT INTO users (user_id, username, password_hash)
-        VALUES (:uid, :u, :ph)
-        """)
-      g.db.execute(query, uid=user_id, u=username, ph='')
-      # Set the password.
-      set_password(username, password)
-      # Invalidate the account creation key.
-      query = text("UPDATE members SET create_account_key = NULL WHERE user_id = :u")
-      g.db.execute(query, u=user_id)
-      trans.commit()
-    except Exception:
-      trans.rollback()
-    # Email the user.
-    query = text("""
-      SELECT CONCAT(fname, ' ', lname) AS name, email
-      FROM members NATURAL JOIN users
-      WHERE username=:u
-      """)
-    result = g.db.execute(query, u=username).first()
-    # Send confirmation email to user.
-    email = result['email']
-    name = result['name']
-    msg = email_templates.CreateAccountSuccessfulEmail.format(name, username)
-    subject = "Thanks for creating an account!"
-    email_utils.sendEmail(email, msg, subject)
-    return True
-  return False
-
