+++ RuddockWebsite/modules/account/__init__.py
@@ -1,0 +1,4 @@
+from flask import Blueprint
+blueprint = Blueprint('account', __name__, template_folder='templates')
+
+import RuddockWebsite.modules.account.routes
