+++ RuddockWebsite/modules/admin/templates/new_members.html
@@ -17,13 +17,13 @@
       </tr>
       <tr>
         <td>Matriculation Year</td>
-        <td><input type="number" name="matriculate_year" /></td>
+        <td><input type="text" name="matriculate_year" /></td>
         <td>Graduation Year</td>
-        <td><input type="number" name="grad_year" /></td>
+        <td><input type="text" name="grad_year" /></td>
       </tr>
       <tr>
         <td>UID</td>
-        <td><input type="number" name="uid" /></td>
+        <td><input type="text" name="uid" /></td>
         <td>Email</td>
         <td><input type="email" name="email" /></td>
       </tr>
